Get dependencies into Docker image without vendoring
Somewhat hacky way whereby we copy the source into the common-build
directory, copy _that_ into the docker image, use it to get the
dependencies (and, importantly, checkpoint), then remove it to allow us
to build against source.

Means that we'll need to `make clean` to update dependencies, but that
shouldn't be too much of a problem, right?