middleware/path_rewrite: Don't inadvertently unescape url-encoded characters (#34)
When a URL is parsed, the original url-encoded path goes in URL.RawPath,
and the decoded path in URL.Path.

URL's string method will use URL.RawPath, preserving the original encoding,
but only if URL.RawPath is a valid encoding of URL.Path.

If we modify URL.Path directly, this is no longer true, and URL then uses the un-escaped version of Path.
This has adverse effects since meaning can differ - for example, if the path contains an encoded / character
(ie. "%2F"), we want to preserve this as an encoded character, as otherwise it appears to be a path seperator.
"/foo%2Fbar" is not the same path as "/foo/bar".

We work around this by applying the regex to the escaped version instead, then reconstructing the unescaped version
from the new escaped version.

The new escaped version SHOULD be a valid url-encoded string - any bad user input would have already caused a 400
during initial parsing. However, there exists a possibility that a regex could produce an invalid result from a valid
input. Since this case is a programmer error, we log.Errorf() and return a 500.